+++ loader/config.txt
@@ -1,4 +1,4 @@
-kernel=baremetal.bin
+kernel=loader.bin
 arm_control=0x200
 kernel_old=1
 disable_commandline_tags=1
