--- loader/printk.h
@@ -1,7 +1,0 @@
-#ifndef PRINTK_H
-#define PRINTK_H
-
-extern void prints(const char *str);
-extern void printk(const char *fmt, ...);
-
-#endif /* PRINTK_H */
