--- loader/timer.h
@@ -1,7 +1,0 @@
-#ifndef TIMER_H
-#define TIMER_H
-
-extern void timer_init();
-extern uint64_t timer_get_syscount();
-
-#endif /* TIMER_H */
